Loot pool, the 2 different versions I've successfully added this is a txt file and won't work if added, please don't add directly to your game
Please let me know if you find a solution here. for now, I will continue to edit without any of the items that have a damage calculation, sticking with hoes, sheild, armour and crossbows and stackable items for now. to note i made 10 attempts but the json could not be read for any that edited the nbt tags.